MeanShift_Classical {MeanShiftR}	R Documentation

Mean shift clustering

Description

Adaptive mean shift clustering to delineate tree crowns from lidar point clouds

Usage

MeanShift_Classical(
  pc,
  H2CW_fac,
  H2CL_fac,
  UniformKernel = FALSE,
  MaxIter = 20L
)

Arguments

pc
Point cloud has to be in matrix format with 3-columns representing X, Y and Z and each row representing one point

H2CW_fac
Factor for the ratio of height to crown width. Determines kernel diameter based on its height above ground.

H2CL_fac
Factor for the ratio of height to crown length. Determines kernel height based on its height above ground.

UniformKernel
Boolean to enable the application of a simple uniform kernel without distance weighting (Default False)

MaxIter
Maximum number of iterations, i.e. steps that the kernel can move for each point. If centroid is not found after all iteration, the last position is assigned as centroid and the processing jumps to the next point

Details

Mean shift clustering

Value

data.frame with X, Y and Z coordinates of each point in the point cloud and X, Y and Z coordinates of the centroid to which the point belongs

[Package MeanShiftR version 0.1.1 Index]