make_CrownPolygons {MeanShiftR}	R Documentation

Fit polygon hulls around tree crowns in a clustered point cloud to derive crown projection areas

Description

The function creates crown projection area polygons from a clustered point cloud of a forest stand. The input point cloud needs to have a column containing tree ID / cluster ID of each point. There are different options for the shape of the polygons.

Usage

make_CrownPolygons(
  pc.dt,
  type = "convexhull",
  N.min = 20,
  Ext.min = 2,
  Ext.max = 50,
  proj4string = CRS(as.character(NA))
)

Arguments

pc.dt
Point cloud in data.table format containing columns X, Y, Z and ID

type
Shape of the desired polygons ("convexhull", "ellipse" or "circle")

N.min
Minimum number of points in a cluster to be considered as a tree crown

Ext.min
Minimum horizontal extent (in meters, in X- and Y-direction) of a cluster to be considered as a tree crown

Ext.max
Maximum horizontal extent (in meters, in X- and Y-direction) of a cluster to be considered as a tree crown

proj4string
Projection string of class CRS-class

Value

SpatialPolygonsDataFrame with each feature representing the crown projection area of one tree and columns containing various geometric attributes

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package MeanShiftR version 0.1.1 Index]