match_CrownsStems {MeanShiftR}	R Documentation

Match crown projection polygons with tree stem positions

Description

The function links remote sensing derived crown projection polygons with inventory derived tree stems. The algorithm sorts all polygons by decreasing area size. Then, starting with the largest crown it checks all stem positions that fall inside the crown projection area and selects the stem with the largest diameter and assigns it to the crown. The stem is removed from the set of available stems and the same procedure is conducted for the next smaller crown and further on until the smallest crown.

Usage

match_CrownsStems(crowns.spdf, stems.spdf, DBH.min = 0.05)

Arguments

crowns.spdf
SpatialPolygonsDataFrame of tree crown projection areas

stems.spdf
SpatialPointsDataFrame of tree stem foot positions (required columns: X, Y, DBH, Species, TreeID)

DBH.min
Minimum stem diameter to be considered

Value

SpatialPolygonsDataFrame of tree crown projection areas with additional columns containing the attributes of the assigned stem for each crown

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package MeanShiftR version 0.1.1 Index]