parallel_MeanShift {MeanShiftR}	R Documentation

Parallel application of mean shift clustering for individual tree crown delineation

Description

The function provides the frame work to apply the adaptive mean shift 3D (AMS3D) algorithm on several sub point clouds of a large investigation area in parallel. It requires a list of sub point clouds as input and returns one large clustered point cloud as output. The input should have buffer zones around the focal areas. The buffer width should correspond to at least the maximal possible tree crown radius.

Usage

parallel_MeanShift(
  pc.list,
  lib.path = NA,
  frac.cores = 0.5,
  version = "classic",
  H2CW = 0.3,
  H2CL = 0.4,
  max.iter = 20,
  buffer.width = 10,
  minz = 2,
  ctr.ac = 2
)

Arguments

pc.list
List of point clouds in data.table format containing columns X, Y and Z (produced by split_BufferedPointCloud function)

lib.path
String specifying the path from where to load the R packages. Should be set to .libPaths()[1].

frac.cores
Fraction of available cores to use for parallelization

version
of the AMS3D algorithm. Can be set to "classic" (slow but precise also with small trees) or "voxel" (fast but based on rounded coordinates of 1-m precision)

H2CW
Factor for the ratio of height to crown width. Determines kernel diameter based on its height above ground.

H2CL
Factor for the ratio of height to crown length. Determines kernel height based on its height above ground.

max.iter
Maximum number of iterations, i.e. steps that the kernel can move for each point. If centroid is not found after all iteration, the last position is assigned as centroid and the processing jumps to the next point

buffer.width
Width of the buffer around the core area in meters

minz
Minimum height above ground for a point to be considered in the analysis. Has to be > 0.

ctr.ac
Centroid accuracy. Specifies the rounding accuracy for centroid positions. After rounding all centroids with the same coordinates are considered to belong to one tree crown.

Value

data.table of point cloud with points labelled with tree IDs

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package MeanShiftR version 0.1.1 Index]