split_BufferedPointCloud {MeanShiftR}	R Documentation

Split point cloud into smaller point clouds with buffer area around

Description

The function splits one large point cloud into several smaller point clouds. It allows to specify a buffer width around the core area where points are included.

Usage

split_BufferedPointCloud(pc.dt, plot.width, buffer.width)

Arguments

pc.dt
Point cloud in data.table format containing columns X, Y and Z

plot.width
Width of the core area in meters

buffer.width
Width of the buffer around the core area in meters

Value

List of sub point clouds and buffer and core points labeled in boolean column "Buffer"

Author(s)

Nikolai Knapp, nikolai.knapp@ufz.de

[Package MeanShiftR version 0.1.1 Index]